Revert "hw7 refactor: queues"
This reverts commit 2ae3462a41fb6336d3a746d44f8b400d6115ed95.
