+++ currency-rates/internal/queue/notification.go
@@ -1,4 +1,4 @@
-package subscriptionnotifier
+package queue
 
 import (
 	"context"
